src/python/requirements.txt
absl-py==1.0.0
anyio==3.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
atomicwrites==1.4.0
attrs==21.4.0
Babel==2.10.1
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
carla-recourse==0.0.4
causalgraphicalmodels==0.0.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.4
cplex==22.1.0.0
cycler==0.11.0
dataclasses==0.6
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
dice-ml==0.5
entrypoints==0.4
fastjsonschema==2.15.3
flake8==4.0.1
fonttools==4.33.2
func-timeout==4.3.5
future==0.18.2
gast==0.5.3
google-pasta==0.2.0
graphviz==0.20
grpcio==1.44.0
h5py==2.10.0
idna==3.3
imageio==2.17.0
importlib-metadata==4.11.3
importlib-resources==5.7.1
iniconfig==1.1.1
ipykernel==6.13.0
ipynb==0.5.1
ipython==7.32.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
jedi==0.18.1
Jinja2==3.1.1
joblib==1.1.0
json5==0.9.6
jsonschema==4.4.0
jupyter-client==7.2.2
jupyter-console==6.4.3
jupyter-core==4.10.0
jupyter-server==1.16.0
jupyterlab==3.3.4
jupyterlab-pygments==0.2.2
jupyterlab-server==2.13.0
jupyterlab-widgets==1.1.0
Keras==2.3.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
kneed==0.7.0
lime==0.2.0.1
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mccabe==0.7.0
mip==1.12.0
mistune==0.8.4
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
nbclassic==0.3.7
nbclient==0.6.0
nbconvert==6.5.0
nbformat==5.3.0
nest-asyncio==1.5.5
networkx==2.6.3
notebook==6.4.11
notebook-shim==0.1.0
numpy==1.19.4
packaging==21.3
pandas==1.1.4
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.0.1
pip==21.2.4
pluggy==1.0.0
prettytable==3.2.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.20.1
psutil==5.9.0
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
Pygments==2.11.2
pyparsing==3.0.8
pyrsistent==0.18.1
pytest==7.1.1
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.3.0
pywin32==303
pywinpty==2.0.5
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.3.0
QtPy==2.0.1
recourse==1.0.0
requests==2.27.1
scikit-image==0.19.2
scikit-learn==0.23.2
scipy==1.7.3
seaborn==0.11.2
Send2Trash==1.8.0
setuptools==61.2.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.13.3
threadpoolctl==3.1.0
tifffile==2021.11.2
tinycss2==1.1.1
tomli==2.0.1
torch==1.7.0
torchvision==0.8.1
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
typing_extensions==4.2.0
urllib3==1.26.9
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.2
Werkzeug==2.1.1
wheel==0.37.1
widgetsnbextension==3.6.0
wincertstore==0.2
wrapt==1.14.0
xgboost==1.4.2
zipp==3.8.0
